[fix] pageでデータを取得して渡すように修正
